Move API routes for get events to one route with url param
